safety use of long double

+++ src/opes/OPESexpanded.cpp
@@ -902,13 +902,13 @@ void OPESexpanded::updateDeltaF(double bias)
 {
   plumed_dbg_massert(counter_>0,"deltaF_ must be initialized");
   counter_++;
-  const double increment=kbt_*std::log1p(std::exp((bias-rct_)/kbt_)/(counter_-1.));
+  const double increment=kbt_*std::log1p(std::exp(static_cast<long double>(bias-rct_)/kbt_)/(counter_-1.));
   #pragma omp parallel num_threads(NumOMP_)
   {
     #pragma omp for
     for(unsigned i=0; i<deltaF_.size(); i++)
     {
-      const double diff_i=(-getExpansion(i)+(bias-rct_+deltaF_[i])/kbt_);
+      const long double diff_i=(-getExpansion(i)+(bias-rct_+deltaF_[i])/kbt_);
       deltaF_[i]+=increment-kbt_*std::log1p(std::exp(diff_i)/(counter_-1.));
     }
   }
